+++ src/main/java/shopping/config/WebMvcConfiguration.java
@@ -5,8 +5,8 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import shopping.config.argumentresolver.AuthenticationPrincipalArgumentResolver;
 import shopping.config.interceptor.AuthInterceptor;
-import shopping.ui.AuthenticationPrincipalArgumentResolver;
 
 import java.util.List;
