+++ src/main/java/shopping/config/argumentresolver/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package shopping.ui;
+package shopping.config.argumentresolver;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -7,6 +7,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import shopping.ui.AuthenticationPrincipal;
 
 @Component
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
